+++ kamehouse-groot/src/main/php/kame-house-groot/api/v1/kamehouse/admin/kamehouse-shell/kamehouse-shell.php
@@ -27,6 +27,8 @@ class KameHouseShell {
     return $kameHouse->util->string->contains($script, "www-data/su.sh") 
       || $kameHouse->util->string->contains($script, "www-data-shell.sh")
       || $kameHouse->util->string->contains($script, "get-kamehouse-secret.sh")
+      || $kameHouse->util->string->contains($script, "encrypt-kamehouse-secrets.sh")
+      || $kameHouse->util->string->contains($script, "decrypt-kamehouse-secrets.sh")
       || $kameHouse->util->string->contains($script, "edit-kamehouse-secrets.sh")
       || $kameHouse->util->string->contains($script, "docker-ssh-")
       || $kameHouse->util->string->contains($script, "ssh.sh");