+++ kamehouse-shell/bin/kamehouse/install-kamehouse-shell.sh
@@ -149,7 +149,7 @@ installKameHouseSecrets() {
   fi
 
   if [ -f "${HOME}/.kamehouse/config/keys/.kamehouse-secrets.cfg" ]; then
-    log.info "WARNING!!!!!!!!!!!!!!!! Found decrypted ${HOME}/.kamehouse/config/keys/.kamehouse-secrets.cfg file. Replace with an encrypted file"
+    log.info "WARNING!!!!!!!!!!!!!!!! Found decrypted ${HOME}/.kamehouse/config/keys/.kamehouse-secrets.cfg file. Encrypt it with encrypt-kamehouse-secrets.sh"
   fi
   chmod -R 700 ${HOME}/.kamehouse/config
 }